requirements.txt
APScheduler>=3.9.1
Flask>=2.0.2
allure-pytest>=2.8.6
pytest>=5.3.2
pytest-dependency>=0.4.0
python-dateutil>=2.8.2
selenium>=4.0.0
webdriver-manager>=3.5.2
chromedriver-autoinstaller>=0.2.2
pyautogui>=0.9.53